+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - (Extractor) Introduce `CharacterModifierInterface` and `DrawableInterface::modify()` to allow modifying drawable objects after extraction
 - (Extractor) Add `Timeline::withAttachment()` and `SpriteDefinition::withAttachment()` to attach a movie clip to a timeline or sprite
 
+**BC breaks:**
+- (Extractor) Remove `FrameObject::characterId` property
+
 0.3.0
 -----